ee.supervisor/requirements.txt
requests
regex
psutil
redis
python-dotenv
kafka-python
pymongo